code/iot-controlled/requirements.txt
rpi_ws281x 
adafruit-circuitpython-neopixel
azure-iot-device
python-dotenv
